Added command to turn laser off in Stop Print
I found 90% of the time when I stop a print the laser is left on. I've added the command the the stop code to send M05 to turn the laser off. Also added a menu item to enable to laser at 1% for visual alignment when jogging. and changed the LCD screen update interval.